Data.ByteArray.Mapping

Description

Documentation

toW64BE :: ByteArrayAccess bs => bs -> Int -> BE Word64

Transform a bytearray at a specific offset into a Word64 tagged as BE (Big Endian)

no bounds checking. unsafe

toW64LE :: ByteArrayAccess bs => bs -> Int -> LE Word64

Transform a bytearray at a specific offset into a Word64 tagged as LE (Little Endian)

no bounds checking. unsafe

fromW64BE :: ByteArray ba => Word64 -> ba

Serialize a Word64 to a ByteArray in big endian format

mapAsWord64 :: ByteArray bs => (Word64 -> Word64) -> bs -> bs

map blocks of 64 bits of a bytearray, creating a new bytestring of equivalent size where each blocks has been mapped through f

no length checking is done. unsafe

mapAsWord128 :: ByteArray bs => (Word128 -> Word128) -> bs -> bs

map blocks of 128 bits of a bytearray, creating a new bytestring of equivalent size where each blocks has been mapped through f

no length checking is done. unsafe